+++ src/engine/Server.cpp
@@ -6,8 +6,6 @@
 
 #include "engine/Server.h"
 
-#include <index/IndexImpl.h>
-
 #include <boost/url.hpp>
 #include <sstream>
 #include <string>
@@ -17,6 +15,7 @@
 #include "engine/ExportQueryExecutionTrees.h"
 #include "engine/QueryPlanner.h"
 #include "global/RuntimeParameters.h"
+#include "index/IndexImpl.h"
 #include "util/AsioHelpers.h"
 #include "util/MemorySize/MemorySize.h"
 #include "util/OnDestructionDontThrowDuringStackUnwinding.h"
@@ -885,27 +884,8 @@ Awaitable<void> Server::processQuery(
   co_return;
 }
 
-struct DeltaTriplesCount {
-  size_t triplesInserted_;
-  size_t triplesDeleted_;
-
-  /// Output as json. The signature of this function is mandated by the json
-  /// library to allow for implicit conversion.
-  friend void to_json(nlohmann::json& j, const DeltaTriplesCount& count) {
-    j = nlohmann::json{
-        {"inserted", count.triplesInserted_},
-        {"deleted", count.triplesDeleted_},
-        {"total", count.triplesInserted_ + count.triplesDeleted_}};
-  }
-
-  DeltaTriplesCount operator-(const DeltaTriplesCount& other) const {
-    return {triplesInserted_ - other.triplesInserted_,
-            triplesDeleted_ - other.triplesDeleted_};
-  }
-};
-
 // ____________________________________________________________________________
-nlohmann::json Server::processUpdateImpl(
+json Server::processUpdateImpl(
     const ad_utility::url_parser::ParamValueMap& params, const string& update,
     ad_utility::Timer& requestTimer, TimeLimit timeLimit, auto& messageSender,
     ad_utility::SharedCancellationHandle cancellationHandle,
@@ -929,12 +909,10 @@ nlohmann::json Server::processUpdateImpl(
                      plannedQuery.parsedQuery_._originalString));
   }
 
-  DeltaTriplesCount countBefore{deltaTriples.numInserted(),
-                                deltaTriples.numDeleted()};
+  DeltaTriplesCount countBefore = deltaTriples.getCounts();
   ExecuteUpdate::executeUpdate(index_, plannedQuery.parsedQuery_, qet,
                                deltaTriples, cancellationHandle);
-  DeltaTriplesCount countAfter{deltaTriples.numInserted(),
-                               deltaTriples.numDeleted()};
+  DeltaTriplesCount countAfter = deltaTriples.getCounts();
 
   LOG(INFO) << "Done processing update"
             << ", total time was " << requestTimer.msecs().count() << " ms"
@@ -947,7 +925,7 @@ nlohmann::json Server::processUpdateImpl(
   // cache key).
   cache_.clearAll();
 
-  nlohmann::json response;
+  json response;
   response["update"] = plannedQuery.parsedQuery_._originalString;
   response["status"] = "OK";
   auto warnings = qet.collectWarnings();
@@ -998,15 +976,14 @@ Awaitable<void> Server::processUpdate(
   auto [cancellationHandle, cancelTimeoutOnDestruction] =
       setupCancellationHandle(messageSender.getQueryId(), timeLimit);
 
-  // Update the delta triples.
-  std::optional<nlohmann::json> response;
-
   // Note: We don't directly `co_await` because of lifetime issues (probably
   // bugs in GCC or Boost) that occur in the Conan build.
   auto coroutine = computeInNewThread(
       updateThreadPool_,
       [this, &params, &update, &requestTimer, &timeLimit, &messageSender,
-       &cancellationHandle, &response] {
+       &cancellationHandle] {
+        std::optional<nlohmann::json> response;
+        // Update the delta triples.
         index_.deltaTriplesManager().modify(
             [this, &params, &update, &requestTimer, &timeLimit, &messageSender,
              &cancellationHandle, &response](auto& deltaTriples) {
@@ -1016,11 +993,11 @@ Awaitable<void> Server::processUpdate(
                   params, update, requestTimer, timeLimit, messageSender,
                   cancellationHandle, deltaTriples);
             });
+        return response;
       },
       cancellationHandle);
-  co_await std::move(coroutine);
+  auto response = co_await std::move(coroutine);
 
-  // TODO<qup42> send a proper response
   // SPARQL 1.1 Protocol 2.2.4 Successful Responses: "The response body of a
   // successful update request is implementation defined."
   AD_CORRECTNESS_CHECK(response.has_value());
